requirements.txt
langchain==0.3.25
langchain-community==0.3.24
langchain-core==0.3.60
langchain-huggingface==0.2.0
langchain-mcp-adapters==0.1.1
langchain-openai==0.3.17
langchain-text-splitters==0.3.8
langgraph==0.4.5
PyPDF2==3.0.1
pypdf==5.5.0
sentence-transformers==4.1.0
fastmcp==2.3.5
a2a-sdk==0.2.8
python-dotenv